improved corporation KillLog.

+++ src/main/java/com/eveonline/api/corporation/KillLog.java
@@ -23,87 +23,188 @@ import java.util.Date;
 
 /**
  * @author Tobias Sarnowski
+ * @author Dominik Eckelmann
  */
 public interface KillLog extends ApiResult {
 
-	int getKillId();
+    /**
+     * Unique kill id.
+     *
+     * You can use this id as beforeKillId in {@link KillLogApi#getKillLogs(com.eveonline.api.DirectorApiKey, long, long)}.
+     * @return unique id of the kill.
+     */
+	long getKillId();
 
-	int getSolarSystem();
+    /**
+     * @return The ID of the solar system the kill occurred in.
+     */
+	long getSolarSystem();
 
+    /**
+     * @return Time the kill occurred.
+     */
 	Date getKillTime();
 
-	int getMoonId();
+    /**
+     * Moon location id.
+     *
+     * Normally populated on a POS los. It is not guaranteed it is present.
+     *
+     * @return Moon location id.
+     */
+	long getMoonId();
 
+    /**
+     * @return get the Victim information.
+     */
 	Victim getVictim();
 
+    /**
+     * @return get Attacker information.
+     */
 	ApiListResult<Attacker> getAttackers();
 
+    /**
+     * @return Item the victim carried.
+     */
 	ApiListResult<Item> getItems();
 
 	/**
 	 * @author Tobias Sarnowski
+     * @author Dominik Eckelmann
 	 */
 	interface Attacker extends ApiResult {
 
-		int getCharacterId();
+        /**
+         * @return ID of the attacker.
+         */
+		long getCharacterId();
 
+        /**
+         * @return Name of the attacker.
+         */
 		String getCharacterName();
 
-		int getCorporationId();
+        /**
+         * @return Corporation id of the attacker's corporation.
+         */
+		long getCorporationId();
 
+        /**
+         * @return Corporation name of the attacker.
+         */
 		String getCorporationName();
 
-		int getAllianceId();
+        /**
+         * @return the alliance id of the attacker.
+         */
+		long getAllianceId();
 
+        /**
+         * @return the attackers alliance name.
+         */
 		String getAllianceName();
 
+        /**
+         * @return security status of the attacker.
+         */
 		float getSecurityStatus();
 
-		int getDamageDone();
+        /**
+         * @return damage done by the attacker.
+         */
+		long getDamageDone();
 
+        /**
+         * @return indicates if he lead the final blow.
+         */
 		boolean hadFinalBlow();
 
-		int getWeaponTypeId();
+        /**
+         * @return Weapon ID the attacker use. Sometimes it can be a ship or missile.
+         */
+		long getWeaponTypeId();
 
-		int getShipTypeId();
+        /**
+         * @return ship of the attacker.
+         */
+		long getShipTypeId();
 
 	}
 
 	/**
 	 * @author Tobias Sarnowski
+     * @author Dominik Eckelmann
 	 */
 	interface Item extends ApiResult {
 
-		int getTypeId();
+        /**
+         * @return the item id.
+         */
+		long getTypeId();
 
+        /**
+         * @return Inventory flags of the item.
+         */
 		int getFlag();
 
-		int getQuantityDropped();
+        /**
+         * @return how much of the items dropped.
+         */
+		long getQuantityDropped();
 
-		int getQuantityDestroyed();
+        /**
+         * @return how much of the item get destroyed.
+         */
+		long getQuantityDestroyed();
 
+        /**
+         * @return get the items that contains other items.
+         */
 		ApiListResult<Item> getContainingItems();
 
 	}
 
 	/**
 	 * @author Tobias Sarnowski
+     * @author Dominik Eckelmann
 	 */
 	interface Victim extends ApiResult {
 
-		int getCharacterId();
+        /**
+         * @return the victim's character id.
+         */
+		long getCharacterId();
 
+        /**
+         * @return the victim's name.
+         */
 		String getCharacterName();
 
-		int getCorporationId();
+        /**
+         * @return the victim's corporation id.
+         */
+		long getCorporationId();
 
+        /**
+         * @return the victim's corporation name.
+         */
 		String getCorporationName();
 
-		int getAllianceId();
+        /**
+         * @return the victim's alliance id.
+         */
+		long getAllianceId();
 
-		int getDamageTaken();
+        /**
+         * @return total ammount of taken damage.
+         */
+		long getDamageTaken();
 
-		int getShipTypeId();
+        /**
+         * @return destroyed ship of the victim.
+         */
+		long getShipTypeId();
 
 	}
 }
